fixes proj image route.. again
